Add docs `Url::class`.
