+++ docs/README.md
@@ -50,6 +50,7 @@ In addition to basic `HTML` tags, the repository also supports the generation of
 This is particularly useful for creating forms and other interactive elements on a webpage.
 
 - [text](/docs/form/input/Text.md)
+- [url](/docs/form/input/Url.md)
 - [week](/docs/form/input/Week.md)
 
 ## Creating Widgets
